Move alphanet testing into a separate integration test, because it interferes with alphanet unit tests
